Doing a little bit of clean up
Moving around some functions to places that make a little more sence.
Also updating headers accordingly.
