+++ Makefile
@@ -6,13 +6,14 @@
 #    By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/08/05 10:49:04 by dpentlan          #+#    #+#              #
-#    Updated: 2023/11/23 16:56:20 by dpentlan         ###   ########.fr        #
+#    Updated: 2023/11/23 22:26:48 by dpentlan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 CC				=	cc
 
 SRC 			=	main/main.c \
+					main/close.c \
 					args/arg_parsing.c \
 					map_parsing/map_data_load.c \
 					map_parsing/map_data_valid.c \
