+++ includes/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 10:46:43 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/23 18:18:43 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/23 22:26:28 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,5 +45,9 @@ typedef struct s_map_data
 	t_rgb		c;
 	char		**map;
 }				t_map_data;
+
+//***** function declaractions *****//
+
+int		ft_close_cub3d(t_clear *clear);
 
 #endif
