+++ srcs/engine/engine.c
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 10:23:46 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/23 18:59:06 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/23 22:27:08 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,29 +16,6 @@
 #include "map_parsing.h"
 #include "mlx.h"
 #include <stdlib.h>
-
-/*
- *	***** ft_close_cub3d *****
- *
- *	DESCRIPTION:
- *		For use with mlx_hook if the X in the window bar is clicked.
- *		Returns int because mlx_hook requires an int return, but function never
- *		returns.
- *		Takes a special t_clear struct which only contains pointers to freeable
- *		structs.
- *	RETURN:
- *		Function shouldn't return. Used with mlx_hook if the x in the window
- *		is pressed.
- *		Also used with mlx_key_hook if ESC key is pressed.
- */
-
-int	ft_close_cub3d(t_clear *clear)
-{
-	ft_free_engine((t_engine *)clear->engine);
-	ft_free_map_data((t_map_data *)clear->map_data);
-	exit(EXIT_SUCCESS);
-	return (0);
-}
 
 /*
  *	***** ft_free_image *****
