+++ srcs/engine/keys.c
@@ -6,12 +6,12 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 12:35:11 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/23 13:47:33 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/23 22:29:52 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
-#include "engine.h"
+#include "cub3D.h"
 
 /*
  *	***** ft_key_press *****
