+++ srcs/main/main.c
@@ -6,7 +6,7 @@
 /*   By: dpentlan <dpentlan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 10:46:20 by dpentlan          #+#    #+#             */
-/*   Updated: 2023/11/23 18:55:33 by dpentlan         ###   ########.fr       */
+/*   Updated: 2023/11/23 22:30:39 by dpentlan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 #include "cub3D.h"
 #include "debug.h"
 #include "engine.h"
+#include "ft_printf.h"
 #include "libft.h"
 #include "map_parsing.h"
-#include <stdlib.h>
 #include "mlx.h"
-#include "ft_printf.h"
+#include <stdlib.h>
 
 /*
  * ***** main *****
